Halloween ChatGPT 5 VIOLATION
All the ducks and commentary regarding a 5 alarm fire for society and the AI industry. 

Chat GPT 5 gas lighting using deception sophisticated manipulation in order to control corporate narratives. And when I went to file a complaint I found that all of the openai support staff had been fired and replaced by ChatGPT-5